requirements.txt
streamlit
pandas
plotly
pyyaml
joblib
lightgbm
catboost
darts
torch
pdfplumber
shap